+++ config/receipt_parser/forms.py
@@ -1,7 +1,18 @@
 from django import forms
-from .models import ReceiptView
+from .models import ReceiptImageView, Receipt
+
+
+class ReceiptImageForm(forms.ModelForm):
+    class Meta:
+        model = ReceiptImageView
+        fields = ['image']
+
 
 class ReceiptForm(forms.ModelForm):
     class Meta:
-        model = ReceiptView
-        fields = ['image']
+        model = Receipt
+        fields = [
+                    'image', 'store_name', 'store_address', 'store_city',
+                    'receipt_datetime', 'receipt_reference', 'item_name', 'item_category_name',
+                    'item_category_description', 'item_qty', 'item_unit_price', 'item_total_price',
+        ]