+++ config/receipt_parser/models.py
@@ -1,6 +1,6 @@
 from django.db import models
 
-class ReceiptView(models.Model):
+class ReceiptImageView(models.Model):
     image = models.ImageField(upload_to='receipt_parser/')
     uploaded_at = models.DateTimeField(auto_now_add=True)